# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface University {
  id: string;
  name: string;
  province: string;
  city: string;
  type: string;
  level: string;
  tags: string[];
  rank_soft?: number;
  rank_qs?: number;
  established?: number;
}

export interface UniversityList {
  items: University[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface Major {
  code: string;
  name: string;
  category: string;
  degree: string;
  years: number;
  employment_rate?: string;
  avg_salary?: string;
}

export interface MajorList {
  items: Major[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
  categories: string[];
}

export interface ControlScore {
  province: string;
  year: number;
  batch: string;
  subjects: string;
  score: number;
}

export interface SchoolScore {
  school_name: string;
  year: number;
  province: string;
  subjects: string;
  batch: string;
  min_score: number;
  min_rank?: number;
}

export type ScoreQueryItemChance =
  (typeof ScoreQueryItemChance)[keyof typeof ScoreQueryItemChance];

export const ScoreQueryItemChance = {
  冲刺: "冲刺",
  稳妥: "稳妥",
  保底: "保底",
} as const;

export interface ScoreQueryItem {
  university: University;
  score: SchoolScore;
  /** 分数差（正数表示超出，负数表示不足） */
  scoreDiff: number;
  chance: ScoreQueryItemChance;
}

export interface ScoreQueryResult {
  items: ScoreQueryItem[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
  inputScore: number;
  province: string;
}

export interface Stats {
  universities: number;
  majors: number;
  provinces: number;
  years: number[];
  count985: number;
  count211: number;
  doubleFirstClass: number;
}

export type ListUniversitiesParams = {
  page?: number;
  pageSize?: number;
  /**
   * 搜索关键词
   */
  search?: string;
  /**
   * 省份筛选
   */
  province?: string;
  /**
   * 院校类型（综合/理工/师范等）
   */
  type?: string;
  /**
   * 标签筛选（985/211/双一流等）
   */
  tag?: string;
  sortBy?: ListUniversitiesSortBy;
};

export type ListUniversitiesSortBy =
  (typeof ListUniversitiesSortBy)[keyof typeof ListUniversitiesSortBy];

export const ListUniversitiesSortBy = {
  rank_soft: "rank_soft",
  rank_qs: "rank_qs",
  name: "name",
  established: "established",
} as const;

export type GetUniversityScoresParams = {
  province?: string;
  year?: number;
};

export type GetUniversityScores200 = {
  scores: SchoolScore[];
};

export type ListMajorsParams = {
  page?: number;
  pageSize?: number;
  search?: string;
  /**
   * 学科门类
   */
  category?: string;
};

export type ListControlScoresParams = {
  province?: string;
  year?: number;
  subjects?: string;
};

export type ListControlScores200 = {
  items: ControlScore[];
};

export type QueryByScoreParams = {
  /**
   * 高考分数
   */
  score: number;
  /**
   * 所在省份
   */
  province: string;
  /**
   * 科类（综合/理科/文科）
   */
  subjects?: string;
  year?: number;
  /**
   * 院校类型筛选
   */
  type?: string;
  /**
   * 标签筛选
   */
  tag?: string;
  page?: number;
  pageSize?: number;
};

export type ListProvinces200 = {
  provinces: string[];
};

export type ListUniversityTypes200 = {
  types: string[];
  tags: string[];
  majorCategories: string[];
};