# lib/api-zod/src/generated/types/controlScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */

export interface ControlScore {
  province: string;
  year: number;
  batch: string;
  subjects: string;
  score: number;
}