# lib/api-zod/src/generated/types/major.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */

export interface Major {
  code: string;
  name: string;
  category: string;
  degree: string;
  years: number;
  employment_rate?: string;
  avg_salary?: string;
}