# lib/api-zod/src/generated/types/majorList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { Major } from "./major";

export interface MajorList {
  items: Major[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
  categories: string[];
}