# lib/api-zod/src/generated/types/scoreQueryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { SchoolScore } from "./schoolScore";
import type { ScoreQueryItemChance } from "./scoreQueryItemChance";
import type { University } from "./university";

export interface ScoreQueryItem {
  university: University;
  score: SchoolScore;
  /** 分数差（正数表示超出，负数表示不足） */
  scoreDiff: number;
  chance: ScoreQueryItemChance;
}