# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  universities: number;
  majors: number;
  provinces: number;
  years: number[];
  count985: number;
  count211: number;
  doubleFirstClass: number;
}