# lib/api-zod/src/generated/types/university.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 高考志愿填报平台 API
 * OpenAPI spec version: 0.1.0
 */

export interface University {
  id: string;
  name: string;
  province: string;
  city: string;
  type: string;
  level: string;
  tags: string[];
  rank_soft?: number;
  rank_qs?: number;
  established?: number;
}